Moved dropdown CSS to their own file
