+++ src/common/components/repository-row/dropdowns/dropdowns.css
@@ -1,0 +1,19 @@
+.helpText {
+  font-size: .875rem;
+  color: $warm-grey;
+}
+
+.buttonRow {
+  padding: 16px;
+  text-align: right;
+}
+
+.cancel {
+  margin:0 2em;
+}
+
+.external {
+  background: url($icon-link-external) 100% 0 no-repeat;
+  background-size: 0.7em;
+  padding-right: 0.9em;
+}
