+++ src/common/components/repository-row/repositoryRow.css
@@ -1,24 +1,4 @@
-.snapName {
+.snapNameInput {
   width: 100%;
   height: 24px;
-}
-
-.helpText {
-  font-size: .875rem;
-  color: $warm-grey;
-}
-
-.buttonRow {
-  padding: 16px;
-  text-align: right;
-}
-
-.cancel {
-  margin:0 2em;
-}
-
-.external {
-  background: url($icon-link-external) 100% 0 no-repeat;
-  background-size: 0.7em;
-  padding-right: 0.9em;
 }
